Fix MSVC warnings that were hidden by q_shared.h.

+++ code/math/Math.h
@@ -33,8 +33,8 @@ Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
 #define M_PI		3.14159265358979323846f	// matches value in gcc v2 math.h
 #endif
 
-#define DEG2RAD( a ) ( ( (a) * M_PI ) / 180.0F )
-#define RAD2DEG( a ) ( ( (a) * 180.0f ) / M_PI )
+#define DEG2RAD( a ) ( ( (a) * (float)M_PI ) / 180.0F )
+#define RAD2DEG( a ) ( ( (a) * 180.0f ) / (float)M_PI )
 
 // angle indexes
 #define	PITCH 0 // up / down
