+++ code/renderer_bgfx/Material.cpp
@@ -33,7 +33,7 @@ int Material::getNumGpuDeforms() const
 {
 	int n = 0;
 
-	for (size_t i = 0; i < numDeforms; i++)
+	for (int i = 0; i < numDeforms; i++)
 	{
 		if (isGpuDeform(deforms[i].deformation))
 			n++;
@@ -44,7 +44,7 @@ int Material::getNumGpuDeforms() const
 
 bool Material::hasCpuDeforms() const
 {
-	for (size_t i = 0; i < numDeforms; i++)
+	for (int i = 0; i < numDeforms; i++)
 	{
 		if (isCpuDeform(deforms[i].deformation))
 			return true;
